Add remapping of entity IDs on deserialize ; make component sets store the whole ID

+++ FoldEngine/Components/ComponentMap.cs
@@ -98,7 +98,7 @@ namespace FoldEngine.Components {
                 //return null;
             }
 
-            return ref ((ComponentSet<T>) Sets[componentType]).Get((int)entityId);
+            return ref ((ComponentSet<T>) Sets[componentType]).Get(entityId);
         }
 
         /// <summary>
